Create Day-11
Given a base- integer, , convert it to binary (base-). Then find and print the base- integer denoting the maximum number of consecutive 's in 's binary representation. When working with different bases, it is common to show the base as a subscript.

Example

The binary representation of 125  is 1111101 . In base 10, there are  5 and 1 consecutive ones in two groups. Print the maximum, 5.