Create Day-13
You are given two classes, Person and Student, where Person is the base class and Student is the derived class. Completed code for Person and a declaration for Student are provided for you in the editor. Observe that Student inherits all the properties of Person.
A char calculate() method that calculates a Student object's average and returns the grade character representative of their calculated average: